+++ example/src/views/CheckedUsed/CheckedUsed.tsx
@@ -1,51 +1,31 @@
 import React, {useCallback, useState} from 'react';
-import {Table, TOnChangePage, IPaginateMeta} from 'bear-react-table';
+import {Table, TPaginateMeta} from 'bear-react-table';
 import styled from 'styled-components/macro';
 import dayjs from 'dayjs';
+import {removeByIndex} from 'bear-jsutils/array';
+
+// Components
+import Checkbox from 'views/_components/Checkbox';
 import {data, IPaginateData} from '../_components/data';
 
 
-const getPageData = (currentPage: number, pageLimit: number, order?: {orderField: string, orderBy: 'DESC'|'ASC'}) => {
 
-    if(order){
-        data.sort((a, b) => mockSort(order.orderBy, order.orderField, a,b));
-    }
-
+const getPageData = (currentPage: number, pageLimit: number) => {
     const pageStart = (currentPage -1) * pageLimit;
     return data.slice(pageStart, pageStart + pageLimit );
 };
 
 
-
-const mockSort = (by: 'DESC'|'ASC', field: string, a: IPaginateData, b: IPaginateData) => {
-
-    const fieldName = field as keyof IPaginateData;
-
-    if (a[fieldName] < b[fieldName]) {
-        return by === 'ASC' ? -1 : 1;
-    }else if (a[fieldName] > b[fieldName]) {
-        return by === 'ASC' ?  1: -1;
-    }
-    // a 必須等於 b
-    return 0;
-};
-
-
-
-const BaseUsed = () => {
-
+const CheckedUsed = () => {
 
     const [isFetching, setIsFetching] = useState(false);
-    const [paginateMeta, setPaginateMeta] = useState<IPaginateMeta>({
+    const [isCheckedAll, setIsCheckAll] = useState<boolean>(false);
+    const [checkedIds, setCheckedIds] = useState<number[]>([]);
+    const [paginateMeta, setPaginateMeta] = useState<TPaginateMeta>({
         currentPage: 1,
         pageLimit: 8,
-        order: {
-            orderField: 'id',
-            orderBy: 'DESC',
-        }
     });
     const [paginateData, setPaginateData] = useState<IPaginateData[]>(getPageData(paginateMeta.currentPage, paginateMeta.pageLimit));
-
     const paginateInfo = {
         totalItems: data.length,
         totalPages: Math.ceil(data.length / paginateMeta.pageLimit),
@@ -53,36 +33,69 @@ const BaseUsed = () => {
 
 
     /**
-     * 查詢分頁
+     * 全選控制
      */
-    const handleFetchPaginate: TOnChangePage = (meta) => {
-        // 取得查詢項目
-        setIsFetching(true);
-        // console.log('meta', meta);
-        setPaginateMeta(meta);
+    const handleCheckedAllId = useCallback((isChecked: boolean) => {
+        setIsCheckAll(isChecked);
+        setCheckedIds(currIds => {
+            if(isChecked){
+                return paginateData.map(row => row.id);
+            }
+            return [];
+        });
 
-        const {currentPage, pageLimit, order} = meta;
+    }, [paginateData]);
 
-        setTimeout(() => {
-            setPaginateData(getPageData(currentPage, pageLimit, order));
-            setIsFetching(false);
-        }, 400);
+
+    /**
+     * 選取單一項目
+     */
+    const handleCheckedId = (isChecked: boolean, id?: string|number) => {
+        const myId = Number(id);
+        setCheckedIds(currIds => {
+            if(!isChecked && currIds.includes(myId)) {
+                const removeId = currIds.findIndex(rowId => rowId === myId);
+                return removeByIndex(currIds, removeId);
+            }
+            return [...currIds, myId];
+        });
     };
 
+    /**
+     * 查詢分頁
+     */
+    const handleFetchPaginate = useCallback((meta: TPaginateMeta) => {
+        // 取得查詢項目
+        setIsFetching(true);
+        setPaginateMeta(meta);
+        setCheckedIds([]);
 
+        setTimeout(() => {
+            setPaginateData(getPageData(meta.currentPage, meta.pageLimit));
+            setIsFetching(false);
+        }, 400);
+    }, []);
+
+
+
+    const handleEdit = (id: number) => {
+    };
 
     return (
         <div>
             <Button type="button" color="primary" onClick={() => setIsFetching(curr => !curr)}>isFetching</Button>
             <div className="d-flex flex-row my-2">
+
+
                 <Table
                     isFetching={isFetching}
                     title={[
-                        {text: '#',          field: 'avatar',      col: 60, titleAlign: 'center', dataAlign: 'center'},
-                        {text: 'Name',       field: 'name',        col: true, isEnableSort: true},
-                        {text: 'Role',       field: 'role',        col: 120},
-                        {text: 'Crated',     field: 'createdAt',   col: 110, isEnableSort: true},
-                        {text: 'Joined',     field: 'isApplyJoin', col: 80},
+                        {text: <Checkbox checked={isCheckedAll} onChange={handleCheckedAllId}/>, field: 'checked', col: 40, titleAlign: 'center', dataAlign: 'center'},
+                        {text: '#', field: 'avatar', col: 60, titleAlign: 'center', dataAlign: 'center'},
+                        {text: 'Name',     field: 'name' , col: true},
+                        {text: 'Role',  field: 'role'        , col: 120},
+                        {text: 'Crated',   field: 'createdAt', col: 110},
+                        {text: 'Joined',      field: 'isApplyJoin', col: 80},
                     ]}
                     data={paginateData.map(row => {
                         const createdAt = dayjs(row.createdAt);
@@ -90,9 +103,9 @@ const BaseUsed = () => {
                         return {
                             id: row.id,
                             disabled: !row.isJoined,
-                            onClickRow: () => console.log(row.id),
                             field: {
                                 role: row.role,
+                                checked: <Checkbox value={row.id} checked={checkedIds.includes(row.id)} onChange={handleCheckedId}/>,
                                 avatar: <Avatar style={{backgroundImage: `url(${row.avatar})`}}/>,
                                 name: <div className="d-flex flex-column">
                                     <div>{row.name}</div>
@@ -102,6 +115,9 @@ const BaseUsed = () => {
                                 createdAt: <div style={{fontSize: 12}}>
                                     {createdAt.format('YYYY-MM-DD')}<br/>
                                     {createdAt.format('HH:mm:ss')}
+                                </div>,
+                                actions: <div style={{fontSize: 12}}>
+                                    <Button color="primary" onClick={() => handleEdit(row.id)}>Edit</Button>
                                 </div>,
                             }
                         };
@@ -121,7 +137,8 @@ const BaseUsed = () => {
 
 };
 
-export default BaseUsed;
+export default CheckedUsed;
+