+++ example/src/views/Router.tsx
@@ -3,6 +3,8 @@ import {Switch, Route, Redirect} from 'react-router-dom';
 
 // Custom
 import BaseUsed from './BaseUsed';
+import CheckedUsed from './CheckedUsed';
+import ExpandUsed from './ExpandUsed';
 
 
 const Router = () => {
@@ -10,6 +12,8 @@ const Router = () => {
     return (
         <Switch>
             <Route path="/" children={<BaseUsed/>} exact/>
+            <Route path="/checkedUsed" children={<CheckedUsed/>} exact/>
+            <Route path="/expandUsed" children={<ExpandUsed/>} exact/>
         </Switch>
     );
 };