+++ src/Table/el-class-names.ts
@@ -21,6 +21,7 @@ export default {
     loadingPosition: `${elClassNames.elCLassPrefix}__loading-position`,
     loadingText: `${elClassNames.elCLassPrefix}__loading-text`,
     loadingImage: `${elClassNames.elCLassPrefix}__loading-image`,
+    icon: `${elClassNames.elCLassPrefix}__icon`,
 
 
     headerSortButton: `${elClassNames.elCLassPrefix}-header__sort-button`,
@@ -39,4 +40,8 @@ export default {
     footerPaginateLi: `${elClassNames.elCLassPrefix}-footer__paginate-li`,
     footerPaginateInfo: `${elClassNames.elCLassPrefix}-footer__paginate-info`,
     footerInner: `${elClassNames.elCLassPrefix}-footer__inner`,
+
+
+    select: `${elClassNames.elCLassPrefix}-footer__select`,
+
 };