+++ src/Table/types.ts
@@ -1,26 +1,31 @@
 import {ReactNode} from 'react';
 
-export interface IData {
-    id: number,
-    [field: string]: any,
-    appendData?: any,
-    disabled?: boolean,
-    onClickRow?: () => void,
+export type sizeUnit = 'px' | '%' | 'em';
+export type TCol = true|number|`${number}${sizeUnit}`;
+
+interface IField {
+    [field: string]: string | number | JSX.Element;
 }
 
-export type TFooterData = React.ReactElement;
+export interface IData {
+    id: number,
+    appendData?: string|number|JSX.Element,
+    disabled?: boolean,
+    onClickRow?: () => void,
+    field: IField
+}
+
+export type TDataFooterContent = React.ReactElement;
 
 export interface ITitle {
+    className?: string,
     field: string,
     text: string|number|ReactNode,
-    col?: number|true,
-    width?: number,
+    col: TCol,
     titleAlign?: 'left'|'center'|'right',
     dataAlign?: 'left'|'center'|'right',
     dataVertical?: 'top'|'center'|'bottom',
-    className?: string,
-    isSort?: boolean,
-    sortBy?: 'DESC'|'ASC',
+    isEnableSort?: boolean,
 }
 
 
@@ -36,15 +41,18 @@ export interface IPaginateInfo {
     totalPages: number,
 }
 
-export interface IPaginateMeta {
+export interface IPage {
     currentPage: number,
-    pageLimit: number,
-    sortField?: string,
-    sortBy?: 'DESC'|'ASC',
+    pageLimit?: number,
+}
+export interface IOrder {
+    orderField: string,
+    orderBy: 'DESC'|'ASC',
 }
 
-
-export interface ICardMeta extends IPaginateMeta {
-    sortField?: string,
-    sortBy?: 'DESC'|'ASC',
+export interface IPaginateMeta extends Required<IPage> {
+    order?: IOrder,
 }
+
+export type TOnChangePage = (meta: IPaginateMeta) => void
+export type TOnChangeSortField = (meta: IOrder) => void